[MY] Further saves some memory

+++ common/src/Engine.c
@@ -29,12 +29,12 @@ UnloadStateCallback unloadStateCallback = NULL;
 
 enum EGameMenuState currentGameMenuState = kResumeCurrentState;
 enum EGameMenuState menuStateToReturn = kResumeCurrentState;
-int cursorPosition = 0;
+int8_t cursorPosition = 0;
 enum EGameMenuState nextNavigationSelection = kResumeCurrentState;
 long timeUntilNextState = MENU_ITEM_TIME_TO_BECOME_ACTIVE_MS;
 enum EPresentationState currentPresentationState;
 size_t biggestOption;
-int isRunning = TRUE;
+uint8_t isRunning = TRUE;
 const char *mainText = NULL;
 int enable3DRendering = TRUE;
 int needsToRedrawHUD = 1;
@@ -120,7 +120,7 @@ void enterState(enum EGameMenuState newState) {
     initialPaintCallback();
 }
 
-int menuTick(long delta_time) {
+uint8_t menuTick(long delta_time) {
 
     enum ECommand input;
     enum EGameMenuState newState;
